Switch from pcap_next_ex to pcap_loop
The reason is that the read timeout of the PCAP handler was used to
check for user interaction, while the documentation explicitly states:

NOTE: the read timeout cannot be used to cause calls that read
packets to return within a limited period of time, because, on
some platforms, the read timeout isn't supported, and, on other
platforms, the timer doesn't start until at least one packet ar-
rives. This means that the read timeout should NOT be used, for
example, in an interactive application to allow the packet cap-
ture loop to ``poll'' for user input periodically, as there's no
guarantee that a call reading packets will return after the
timeout expires even if no packets have arrived.

So now pcap_loop is used, but this required a whole rework of the loop
and the dispatcher thread. Now the main thread handles SIGINT and
SIGTERM calling pcap_breakloop, while the dispatcher SIGUSR1 and SIGALRM
only. Also the alarm is sent always (not only for active killing mode)
so that the dispatcher is periodically woken up and can check for the
termination flag, which is set from the main thread when the pcap_loop
function terminates. The reason for that is also that there is (AFAIK) a
portable version of sigtimedwait. The only drawback should be that the
termination is delayed at most of DISPATCHER_TIMEOUT (1 second).